Moved socket Events into startup file
